pyproject.toml
[tool.poetry]
name = "application"
version = "0.1.0"
description = ""
authors = ["Hazel Lynn James"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.1"
python-dotenv = "^0.20.0"
mysql-connector-python = "^8.0.28"
Flask-Admin = "^1.6.0"
sqlacodegen = "^2.3.0"
Flask-Login = "^0.6.0"
Flask-WTF = "^1.0.1"
WTForms = {extras = ["email"], version = "^3.0.1"}
email-validator = "^1.1.3"
Flask-Breadcrumbs = "^0.5.1"
jupyterlab = "^3.4.0"
WTForms-SQLAlchemy = "^0.3"
black = "^22.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
